Stealth cameras and lights for soldiers
Big change and lots of stuff changed, todo with more stuff included in this commit and that will be a full stealth revamp essentially, and hopefully provide some more infra to add more enemies - but maybe not. ideally we make the FOV its own thing that's easily shared yet individual across all the stuff... not sure how to do this. maybe  BaseStealthEnemy with a fov and stuff? and then the actual fov code checks against that. just rambling now.